backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.0
sentence-transformers>=2.6.0
faiss-cpu>=1.8.0
google-generativeai>=0.4.1
firebase-admin>=6.5.0
pillow>=10.3.0
python-dotenv>=1.0.1
requests>=2.31.0
protobuf==4.25.3
duckduckgo-search>=5.0.0
gunicorn>=21.2.0
